添加移动账号

+++ src/pages/category-list/reducer.ts
@@ -1,11 +1,16 @@
 import categoryApi from '@/apis/categoryApi';
-import { showToast } from '@/common/utils';
+import { showConfirm, showToast } from '@/common/utils';
 import { AppDispatch, RootState } from '@/redux/store';
 import { createSlice } from '@reduxjs/toolkit';
 
 // 为 slice state 定义一个类型
 interface CategoryListState {
   list: CategoryInfo[]; // 分类列表
+
+  // more actionsheet
+  moreOpened: boolean;
+  moreItem?: CategoryInfo;
+  moreItemIndex?: number;
 
   // 数据已经变化，需要刷新
   dataChanged: boolean;
@@ -14,6 +19,11 @@ interface CategoryListState {
 // 使用该类型定义初始 state
 const initialState: CategoryListState = {
   list: [],
+
+  moreOpened: false,
+  moreItem: undefined,
+  moreItemIndex: undefined,
+
   dataChanged: false,
 };
 
@@ -47,22 +57,34 @@ export const categoryListSlice = createSlice({
       state.dataChanged = true;
     },
 
-    // 关闭列表项的SwipeAction
-    setSwipeActionOpened: (state, action) => {
-      for (let i = 0; i < state.list.length; i++) {
-        const item = state.list[i];
-        item.isOpened = i === action.payload;
+    // 删除分类
+    deleteCategory: (state, action) => {
+      let id = action.payload;
+      let index = state.list.findIndex((item) => item.id === id);
+      if (index >= 0) {
+        state.list.splice(index, 1);
       }
     },
-    // 打开列表项的SwipeAction
-    setSwipeActionClosed: (state, action) => {
-      let index = action.payload;
-      state.list[index].isOpened = false;
+
+    // 打开MoreActionSheet
+    openMoreActionSheet: (state, action) => {
+      const { item, index } = action.payload;
+      state.moreOpened = true;
+      state.moreItem = item;
+      state.moreItemIndex = index;
+    },
+
+    // 关闭MoreActionSheet
+    closeMoreActionSheet: (state) => {
+      state.moreOpened = false;
+      state.moreItem = undefined;
+      state.moreItemIndex = undefined;
     },
   },
 });
 
-export const { setCategoryListState, clearCategoryListState, setCategoryList, setCategoryListDataChanged, setSwipeActionOpened, setSwipeActionClosed } = categoryListSlice.actions;
+export const { setCategoryListState, clearCategoryListState, setCategoryList, setCategoryListDataChanged, deleteCategory, openMoreActionSheet, closeMoreActionSheet } =
+  categoryListSlice.actions;
 export const categoryListSelector = (state: RootState) => state.categoryList;
 
 /**
@@ -76,9 +98,20 @@ export const fetchCategoryList = () => async (dispatch: AppDispatch) => {
 /**
  * 删除分类
  */
-export const deleteCategoryAsync = (id: number, index?: number) => async (dispatch: AppDispatch) => {
-  await categoryApi.deleteCategory({ id });
-  showToast('删除成功');
+export const deleteCategoryAsync = (id: number) => async (dispatch: AppDispatch) => {
+  await new Promise((resolve, reject) => {
+    showConfirm({
+      content: '您确定要删除该分类吗？',
+      onOk: async () => {
+        await categoryApi.deleteCategory({ id });
+        await showToast('删除成功');
+
+        dispatch(deleteCategory(id));
+
+        resolve(true);
+      },
+    });
+  });
 };
 
 export default categoryListSlice.reducer;
